feat(logtime): add logtime feature with calendar and statistics display
- Implemented logtime.css for styling the logtime feature.
- Created logtime.ts to handle logtime data fetching, rendering, and configuration.
- Added functions to create a calendar grid and month cards displaying logged time.
- Integrated taco count display and goal tracking with visual indicators.
- Enhanced user interaction with a scrollable logtime display.

feat(profile): enhance profile page with event filtering and customization

- Introduced event filtering based on campus and event type in events.ts.
- Added profile.ts to manage profile customization including avatar, banner, and background images.
- Implemented a dropdown for event type filtering with persistent settings.
- Ensured UI updates dynamically with MutationObserver for real-time changes.
