+++ package.json
@@ -1,5 +1,5 @@
 {
-  "name": "42_intra_hub",
+  "name": "better-intra",
   "private": true,
   "version": "0.0.1",
   "type": "module",
@@ -9,9 +9,9 @@
     "preview": "vite preview"
   },
   "devDependencies": {
+    "@awesome.me/webawesome": "^3.6.0",
     "typescript": "~6.0.2",
     "vite": "^7.3.2",
-    "vite-plugin-monkey": "^7.1.9",
-    "@awesome.me/webawesome": "^3.5.0"
+    "vite-plugin-monkey": "^7.1.9"
   }
 }
