+++ src/features/logtime/logtime.css
@@ -1,0 +1,206 @@
+.lt-box-container {
+  font-family: var(--intra-font);
+  margin-bottom: 0 !important;
+}
+[class*="logtime"],
+.card,
+.card-body,
+.row,
+.col-md-12 {
+  overflow: visible !important;
+  height: auto !important;
+  min-height: 0 !important;
+  margin: 0 !important;
+  padding: 0 !important;
+}
+.log-slider-fixed {
+  overflow-x: auto !important;
+  overflow-y: hidden !important;
+  max-height: fit-content !important;
+  align-items: flex-start !important;
+  width: 100%;
+  margin-top: -10px !important;
+  display: flex !important;
+  scroll-behavior: smooth;
+  cursor: grab;
+  user-select: none;
+  scrollbar-width: none;
+}
+.log-slider-fixed::-webkit-scrollbar {
+  display: none !important;
+}
+.grid-centering-container {
+  display: flex;
+  gap: 16px;
+  padding: 0 10px !important;
+  margin: 0 auto;
+  min-width: min-content;
+}
+.month-card {
+  transition: opacity 0.2s ease;
+  border: 0.5px solid var(--border-color);
+  background: "#ffffff";
+  border-radius: 16px;
+  padding: 10px;
+  width: 280px;
+  flex-shrink: 0;
+  box-shadow: 0 4px 6px -1px rgba(0, 0, 0, 0.05);
+  display: flex;
+  flex-direction: column;
+}
+.month-card.current-month {
+  border: 3px solid var(--calendar-color) !important;
+  box-shadow: 0 0 15px var(--calendar-color) 33 !important;
+  opacity: 1 !important;
+}
+.day-cell {
+  transition: transform 0.1s ease;
+  cursor: pointer;
+  position: relative;
+}
+.day-cell:hover {
+  transform: scale(1.2);
+  filter: brightness(0.9);
+  z-index: 50 !important;
+}
+.today-highlight {
+  box-shadow:
+    0 0 0 2px rgba(239, 68, 68, 0.65),
+    0 0 12px rgba(239, 68, 68, 0.45);
+  border: none !important;
+  z-index: 10;
+}
+.day-tooltip {
+  display: none;
+  position: absolute;
+  bottom: 130%;
+  left: 50%;
+  transform: translateX(-50%);
+  background: #1e293b;
+  color: #fff;
+  font-size: 11px;
+  padding: 4px 8px;
+  border-radius: 4px;
+  z-index: 100;
+  white-space: nowrap;
+  pointer-events: none;
+}
+.day-cell:hover .day-tooltip {
+  display: block;
+}
+@keyframes rainbow-shift {
+  0% {
+    background-position: 0% 50%;
+  }
+  50% {
+    background-position: 100% 50%;
+  }
+  100% {
+    background-position: 0% 50%;
+  }
+}
+.badge-rainbow {
+  background: linear-gradient(
+    270deg,
+    #ff0000,
+    #ff7f00,
+    #ffeb3b,
+    #4caf50,
+    #00bcd4,
+    #2196f3,
+    #3f51b5,
+    #9c27b0,
+    #e91e63,
+    #ff0000
+  );
+  background-size: 400% 400% !important;
+  animation: rainbow-shift 3s ease infinite !important;
+  border: none !important;
+  color: white !important;
+  text-shadow: 0px 1px 2px rgba(0, 0, 0, 0.2);
+}
+@keyframes wave-move {
+  0% {
+    transform: rotate(0deg);
+  }
+  100% {
+    transform: rotate(360deg);
+  }
+}
+
+.liquid-container {
+  width: 100%;
+  height: 12px;
+  background: #f1f5f9;
+  border-radius: 10px;
+  margin-bottom: 16px;
+  overflow: hidden;
+  position: relative;
+  border: 1px solid rgba(0, 0, 0, 0.03);
+}
+
+.liquid-fill {
+  height: 100%;
+  background: var(--calendar-color);
+  transition: width 1s ease-in-out;
+  position: relative;
+}
+
+.liquid-fill::after {
+  content: "";
+  position: absolute;
+  top: -12px;
+  right: -20px;
+  width: 40px;
+  height: 40px;
+  background: rgba(255, 255, 255, 0.4);
+  border-radius: 38%;
+  animation: wave-move 3s linear infinite;
+  display: block;
+}
+
+.liquid-fill-full {
+  height: 100%;
+  background: linear-gradient(270deg, #ff0000, #ffeb3b, #e91e63, #ff0000);
+  background-size: 400% 400%;
+  animation: rainbow-shift 3s ease infinite;
+  width: 100%;
+}
+@keyframes taco-bounce {
+  0%,
+  100% {
+    transform: translateY(0);
+  }
+  50% {
+    transform: translateY(-3px);
+  }
+}
+
+.taco-bank {
+  display: flex;
+  align-items: center;
+  gap: 8px;
+  background: rgba(255, 165, 0, 0.1);
+  padding: 4px 12px;
+  border-radius: 20px;
+  border: 1px solid rgba(255, 165, 0, 0.3);
+  margin-left: 10px;
+  transition: all 0.3s ease;
+  cursor: help;
+}
+
+.taco-bank:hover {
+  background: rgba(255, 165, 0, 0.2);
+  transform: scale(1.05);
+}
+
+.taco-icon {
+  animation: taco-bounce 2s infinite ease-in-out;
+  font-size: 16px;
+}
+
+.taco-count {
+  font-weight: 800;
+  color: #d97706;
+  font-size: 14px;
+}
