+++ src/features/vite-env.d.ts
@@ -2,6 +2,10 @@ declare module "*.css" {
   const content: any;
   export default content;
 }
+declare module "*.css?inline" {
+  const content: any;
+  export default content;
+}
 declare module "*.svg?raw" {
   const content: string;
   export default content;
